Seperated persistence from PlayBackController
